backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
fpdf
Pillow
requests
python-dotenv
openai